Some changes to the output and added label to the resulting tree
Now the resulting tree will have three values. The first is a label
representing the constraint this was from the analysis (allowing user
to check the second best trees for each branch in the constraint output
files).
Second, it the difference in log likelihood between the best tree and
the best reverse constraint for this branch.
Third, is the proportion of sites in the alignment that support the
branch from the best tree over the second best arrangement (form the
constraint analysis).