Added bootstrap analysis
Machete can now carry out a bootstrap analysis using the best models
found for the data. This is separate to the reverse constraints
analysis which only uses the ML tree to identify the constraints to be
analysed.
Input options have been added to start the bootstrap “-r INT” where INT
is the number of bootstrap reps to be carried out
“-n” which tells machete NOT to carry out the reverse constraints
analysis (and only build the best tree).